+++ replication/src/messages/mod.rs
@@ -1,0 +1,5 @@
+use std::marker::PhantomData;
+use crate::prelude::*;
+
+#[derive(Default)]
+pub(crate) struct ReplicationData<T: Replicable>(PhantomData<T>);
